global_planner: use default_tolerance parameter
The default_tolerance parameter did not seem to be used by the
global_planner.
If a goal is set inside an obstacle, the global planner tries to find a
valid cell in the backwards direction given by the goal until the length
defined by default_tolerance.
This behaviour is basically taken from the carrot_planner.