global_planner: always perform potential field
We simply always calculate the potential field regardless of if we're
in a recovery behaviour or not.
This saves time as the potential field method does not break when the
goal is free.

We also only re-calculate the potential field if the last selected goal
from the potential field method becomes invalid.