global_planner: set last_goal to current goal on init
The last_goal being [0,0] for the potential field search
method caused errors when the origin of the map is at [0,0] (segfault).
This commit makes sure that the "last_goal" is set to the actual goal
pose (as passed to the global planner). When the actual goal pose is a
valid goal we do not have to make use of the potential field method and
just go ahead with the actual goal.