global_planner: enable default_tolerance with potential field
Previously the default_tolerance parameter was not used leading to
failure when the desired goal was inside an obstacle.
In case the goal is inside an obstacle the global_planner now tries to
find a cell with pot_field_max_cost as its maximum value as close as
possible around the desired goal pose. This is achieved by calculating a
potential field in the area of the desired goal pose.
This area is defined by the default_tolerance parameter.